First cut at new directory-watch option ('-d').
This mode will watch directories implicitly

    ls -d src | entr -d ./build

Or implicitly based on the dirname(3) of each file provided as input

    ls src/* | entr -d ./build

Multiple references to the same directory are deduped

Since `entr` relies on external tools to find files we rely on an external shell loop to react to an altered directory. If a file is added to a watch directory it will gracefully exit. Standard usage is as follows

    while :; do
       ls src/*.rb | entr -d rake
    done

System tests pass on OpenBSD 5.5, MacOS 10.7, CentOS 6.5